Distinguish leaf nodes from intermediary nodes of the same name
Vault's API will allow for intermediary nodes to be named
identically to leaf nodes. When it returns data in the listing
API, you will see `node1/` and `node1` as separate items.

In order to properly display, copy, and move these, `safe`
now makes a distinction between `node1/` and `node`. This
will show up in `safe tree` as a blue directory node, and a
green leaf node.

This fixes a bug in `safe tree` and `safe paths` whereby
the listings of `node1/` would be duplicated, and the listing
of `node1` was overlooked.

This also fixes a bug in the recursive move/copy, allowing
for it to properly copy/move leaf nodes that were also named
as intermediary nodes, and allows us to skip copying/moving empty
top-level nodes (which used to result in an error to the user).