Recursive delete (#33)
* Add support for recursive move/copy/delete

Users are now able to move, copy, or delete
trees in Vault recursively, via the `-R` flag.
This will prompt for confirmation. If you wish
to ignore confirmation and copy/move/delete blindly,
use the `-f` flag as well.

NOTE: Vault does not appear to delete empty directories,
which may lead to confusion after a move/delete, and
the directories still appear. Any keys/values inside
those directories have been removed.

* Added request debugging

* Distinguish leaf nodes from intermediary nodes of the same name

Vault's API will allow for intermediary nodes to be named
identically to leaf nodes. When it returns data in the listing
API, you will see `node1/` and `node1` as separate items.

In order to properly display, copy, and move these, `safe`
now makes a distinction between `node1/` and `node`. This
will show up in `safe tree` as a blue directory node, and a
green leaf node.

This fixes a bug in `safe tree` and `safe paths` whereby
the listings of `node1/` would be duplicated, and the listing
of `node1` was overlooked.

This also fixes a bug in the recursive move/copy, allowing
for it to properly copy/move leaf nodes that were also named
as intermediary nodes, and allows us to skip copying/moving empty
top-level nodes (which used to result in an error to the user).

* `tree` and `paths` will not show paths to nowhere

`safe tree` and `safe paths` will now hide any intermediary
nodes that do not have any leaf nodes underneath them (or any
of their child-intermediary nodes).

This works around the inability to clean up intermediary node
references. Recursive move/copy/delete are unaffected,
as they only interact with leaf nodes and leaf nodes
are distinct from intermediary nodes of the same name.