Implement seal/unseal/status ops for HA vaults
*NOTE* all of this functionality relies on strongbox to determine safely
which Vaults are sealed and which are unsealed.  You can find the code
for that here: https://github.com/jhunt/go-strongbox

`safe seal` will interrogate strongbox to find all unsealed vaults,
and then seal those vaults transparently, waiting for background Consul
leader elections to take place.

`safe status` basically queries strongbox and pretty-prints what it
finds.  This is surprisingly useful.

`safe unseal` will find the sealed vaults, and determine how many seal
keys are required for an unseal.  It then prompts for the correct
number (up front) and then re-unseals all sealed backends.  It properly
resets seal progress on each backend first, to handle half-unsealed
Vaults.

This is an alternate implementation to PR #45, which is technically less
flexible (requires a helper component on the server deployment) but is
muuuuuuch simpler.