Fix tests / colorization-related bugs
`safe tree` behaved differently when run without colorization.  This
caused the tests to begin failing when we updated to jhunt/go-ansi,
which properly detects _output_ terminal, not stdin (as old goutils did)

When this was fixed, the paths / export commands, which were relying on
this, started failing tests due to doubled-up slashes, so we added a new
StripSlashes flag to TreeOptions to capture this unintended side-effect
of UseANSI: false.