Remove broken tests; revert CI to `make test`
The pipeline update broke our test suite, because it assumed we wanted
to use `go test` and `go vet`, which we most certainly do not.
Additionally, chased a bit of a red herring since we _have_ go tests
defined (apparently) but they don't work and never have. I removed them.