Fix ls mount listing
First thing is that the mount listing was looking for mounts of type
secret, which is not, nor to my knowledge has ever been, a thing. So
I changed that to "generic". Also, there was a formatting error where
on newer versions of Vault, mounts would be listed with a trailing
double slash which, while innocuous, is effectively more annoying
than the other bug where mounts weren't all listing in the first place.