x509 validate works with openssl generated keys
openssl req makes PKCS#8 private keys, and so trying to validate
them would lead to - first - a complaint that the PEM header wasn't
valid, and then - second - that it could not parse the key (because
it only tries to parse a PKCS#1). Now validate can handle PKCS#8
with a "BEGIN PRIVATE KEY" header.