Multithread tree (#170)
* Multithreaded tree, paths, export calls

This rewrites the core code that drives tree, paths, and export.
Requests to the vault are handled concurrently by worker threads,
showing a significant speedup in requests that had previously taken a
long period of time (and typically a minor speedup in those that did
not). As an added bonus, this also fixes a long-existent oddity where
safe did not error when trying to tree or paths a non-existent
path. Also, now you can specify "/" as a path to tree, paths, or export
to get the entire vault kv tree - not just the secrets mount.
There is a minor formatting difference in tree, such that the requested
path will appear as a directory if it is only a path (not a secret).
This seems like an improvement, but it is a behavioral difference nonetheless.

* recursive copy, move, and delete use new tree core

Deleted old Tree code because nothing is using it anymore.