Implement deep copy and deep move
This required a rework of the concurrency tree code. Turns out, for most
purposes, also, walking a tree is not the most convenient way to access
this data for most operations. So now its basically in a slice. Bunch of
bug fixes for the tree code. Move still isn't working _quite_ right.