Fix tree recursion when items are deleted
This is a small reworking of how deleted/destroyed secrets are handled
when walking the tree. This fixes a few things - rm won't bomb when a
recursive delete is done with the mount as the root. Recursive move can
delete the tree properly. Deleting the entire mount tree works properly,
whether there are already deleted items or not.