make safe paths have output again
We weren't pruning the version information when getting paths, and so
printing the paths was walking the versions and finding no keys.
Therefore it was printing nothing. Also, there was a bug in the tree
code where if something was a secret and a directory, it would not
retrieve the versions in all the right circumstances.