tree/paths works when secret is at root of mount
If you have mount "secret" and then safe set secret, it would work. But
the way Vault outputs a LIST is different at the root so we had to
ignore the secret output of the mount list so that we don't duplicate
the secret in the output (which would then panic tree because tree
rightly assumes there are no duplicate paths).