Bump vaultkv to work with older versions of vault
Versions of vault before 0.9 where the user had a token that could not
read sys endpoints would result in a 403 propagating back to the user.
We now capture the 403, check to see if the users token is valid, and
then if so, determine that the backend must be v1 (or non-existent).
From there, things should work as expected.