fix safe seal on a clustered Vault
safe was bailing out when it tried to seal a Vault in standby mode
(because you can't). Because safe tries to seal _all_ the Vaults in a
cluster, it needs to strategically only seal the master at any given
time, and then wait for a new master to be elected, then seal that,
until there's nothing left unsealed. The existing code would work if the
Vaults became master in IP sequential order, but that's it.