vaultkv reuses connections now. Faster tree output
vaultkv was reading the response body contents up until the byte before
the EOF, but not actually triggering the EOF flag and allowing the
Client to know it can reuse the connection. This was leading to commands
like safe tree to perform TCP/TLS handshakes potentially thousands of
times. Now it only needs to do it once for each thread that is doing work.
The speed increase can be significant when running against large Vaults
with moderate latency in the connection.