Add CA cert flag to target
Passes cert through to vault CLI by writing a temp file and exposing the
path of it to Vault through the VAULT_CACERT envvar. It should get
cleaned up in the cases of normal shutdown and SIGINT being sent to the
program.

Some of the connection configuration logic got moved out of the vault
package and into main. NewVault is now configured with a configuration
struct instead of reading environment variables directly.